Move controls into navbar
